+++ semana17/boilerplate_autenticacao/.gitignore
@@ -2,3 +2,4 @@ node_modules
 package-lock.json
 build
 .env
+READMEaula50.md
