--- semana17/boilerplate_autenticacao/src/services/generateToken.ts
@@ -1,14 +1,0 @@
-import * as jwt from "jsonwebtoken";
-import { AuthenticationData } from "../types";
-
-const expiresIn = "1min";
-
-export const generateToken = (input: AuthenticationData): string => {
-    return jwt.sign(
-        input,
-        process.env.JWT_KEY as string,
-        {
-            expiresIn: process.env.JWT_EXPIRES_IN
-        }
-    )
-}
