+++ semana17/boilerplate_autenticacao/src/services/hashManager.ts
@@ -1,0 +1,14 @@
+import * as bcrypt from "bcryptjs";
+
+
+export const hash = async(s: string): Promise<string> => {
+    const rounds = Number(process.env.BCRYPT_COST);
+    const salt = await bcrypt.genSalt(rounds);
+    const result = await bcrypt.hash(s, salt);
+    return result;
+}
+
+export const compare = async(s: string, hash: string): Promise<boolean> => {
+    return bcrypt.compare(s, hash);
+}
+
