+++ semana17/boilerplate_autenticacao/src/types.ts
@@ -5,5 +5,5 @@ export type AuthenticationData = {
 export type User = {
     id: string,
     email: string,
-    password: string
+    password: string,
 }
